add quota group
