+++ src/Areas/Deploy/Commands/DeployJsonContext.cs
@@ -2,8 +2,6 @@
 // Licensed under the MIT License.
 
 using System.Text.Json.Serialization;
-using Areas.Deploy.Services.Util;
-using AzureMcp.Areas.Deploy.Commands.Quota;
 using AzureMcp.Areas.Deploy.Commands.Region;
 using AzureMcp.Areas.Deploy.Models;
 using AzureMcp.Areas.Deploy.Options;
@@ -18,9 +16,6 @@ namespace AzureMcp.Areas.Deploy.Commands;
 [JsonSerializable(typeof(AppTopology))]
 [JsonSerializable(typeof(MermaidData))]
 [JsonSerializable(typeof(MermaidConfig))]
-[JsonSerializable(typeof(QuotaCheckCommand.QuotaCheckCommandResult))]
-[JsonSerializable(typeof(QuotaInfo))]
-[JsonSerializable(typeof(Dictionary<string, List<QuotaInfo>>))]
 [JsonSerializable(typeof(RegionCheckCommand.RegionCheckCommandResult))]
 [JsonSerializable(typeof(List<string>))]
 internal sealed partial class DeployJsonContext : JsonSerializerContext
