+++ src/Areas/Deploy/Models/EncodeMermaid.cs
