+++ src/Areas/Deploy/Services/IDeployService.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using Areas.Deploy.Services.Util;
 using AzureMcp.Areas.Deploy.Models;
 using AzureMcp.Options;
 
@@ -14,12 +13,6 @@ public interface IDeployService
         string azdEnvName,
         string subscriptionId,
         int? limit = null);
-
-    Task<Dictionary<string, List<QuotaInfo>>> GetAzureQuotaAsync(
-        List<string> resourceTypes,
-        string subscriptionId,
-        string location);
-
 
     Task<List<string>> GetAvailableRegionsForResourceTypesAsync(
         string[] resourceTypes,
