+++ src/Areas/Quota/Commands/QuotaCheckCommand.cs
@@ -1,24 +1,24 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using Areas.Deploy.Services.Util;
-using AzureMcp.Areas.Deploy.Options;
-using AzureMcp.Areas.Deploy.Services;
+using AzureMcp.Areas.Quota.Services.Util;
+using AzureMcp.Areas.Quota.Options;
+using AzureMcp.Areas.Quota.Services;
 using AzureMcp.Commands;
 using AzureMcp.Commands.Subscription;
 using AzureMcp.Models.Command;
 using AzureMcp.Services.Telemetry;
 using Microsoft.Extensions.Logging;
 
-namespace AzureMcp.Areas.Deploy.Commands.Quota;
+namespace AzureMcp.Areas.Quota.Commands;
 
 public class QuotaCheckCommand(ILogger<QuotaCheckCommand> logger) : SubscriptionCommand<QuotaCheckOptions>()
 {
     private const string CommandTitle = "Check Available Azure Quota for Regions";
     private readonly ILogger<QuotaCheckCommand> _logger = logger;
 
-    private readonly Option<string> _regionOption = DeployOptionDefinitions.QuotaCheck.Region;
-    private readonly Option<string> _resourceTypesOption = DeployOptionDefinitions.QuotaCheck.ResourceTypes;
+    private readonly Option<string> _regionOption = QuotaOptionDefinitions.QuotaCheck.Region;
+    private readonly Option<string> _resourceTypesOption = QuotaOptionDefinitions.QuotaCheck.ResourceTypes;
 
     public override string Name => "quota-check";
 
@@ -64,8 +64,8 @@ public class QuotaCheckCommand(ILogger<QuotaCheckCommand> logger) : Subscription
                 .Select(rt => rt.Trim())
                 .Where(rt => !string.IsNullOrWhiteSpace(rt))
                 .ToList();
-            var deployService = context.GetService<IDeployService>();
-            Dictionary<string, List<QuotaInfo>> toolResult = await deployService.GetAzureQuotaAsync(
+            var quotaService = context.GetService<IQuotaService>();
+            Dictionary<string, List<QuotaInfo>> toolResult = await quotaService.GetAzureQuotaAsync(
                 ResourceTypes,
                 options.Subscription!,
                 options.Region);
@@ -75,7 +75,7 @@ public class QuotaCheckCommand(ILogger<QuotaCheckCommand> logger) : Subscription
             context.Response.Results = toolResult?.Count > 0 ?
                 ResponseResult.Create(
                     new QuotaCheckCommandResult(toolResult),
-                    DeployJsonContext.Default.QuotaCheckCommandResult) :
+                    QuotaJsonContext.Default.QuotaCheckCommandResult) :
                 null;
         }
         catch (Exception ex)
