+++ src/Areas/Quota/Services/Util/HDInsightQuotaChecker.cs
@@ -2,7 +2,7 @@ using Azure.Core;
 using Azure.ResourceManager.HDInsight;
 using Azure.ResourceManager.HDInsight.Models;
 
-namespace Areas.Deploy.Services.Util;
+namespace AzureMcp.Areas.Quota.Services.Util;
 
 public class HDInsightQuotaChecker(TokenCredential credential, string subscriptionId) : AzureQuotaChecker(credential, subscriptionId)
 {
