+++ src/Areas/Quota/Services/Util/StorageQuotaChecker.cs
@@ -1,20 +1,21 @@
 using Azure.Core;
-using Azure.ResourceManager.Search;
-using Azure.ResourceManager.Search.Models;
+using Azure.ResourceManager;
+using Azure.ResourceManager.Storage;
+using Azure.ResourceManager.Storage.Models;
 
-namespace Areas.Deploy.Services.Util;
+namespace AzureMcp.Areas.Quota.Services.Util;
 
-public class SearchQuotaChecker(TokenCredential credential, string subscriptionId) : AzureQuotaChecker(credential, subscriptionId)
+public class StorageQuotaChecker(TokenCredential credential, string subscriptionId) : AzureQuotaChecker(credential, subscriptionId)
 {
     public override async Task<List<QuotaInfo>> GetQuotaForLocationAsync(string location)
     {
         try
         {
             var subscription = ResourceClient.GetSubscriptionResource(new ResourceIdentifier($"/subscriptions/{SubscriptionId}"));
-            var usages = subscription.GetUsagesBySubscriptionAsync(location);
+            var usages = subscription.GetUsagesByLocationAsync(location);
             var result = new List<QuotaInfo>();
 
-            await foreach (QuotaUsageResult item in usages)
+            await foreach (var item in usages)
             {
                 result.Add(new QuotaInfo(
                     Name: item.Name?.Value ?? string.Empty,
@@ -28,7 +29,7 @@ public class SearchQuotaChecker(TokenCredential credential, string subscriptionI
         }
         catch (Exception error)
         {
-            Console.WriteLine($"Error fetching Search quotas: {error.Message}");
+            Console.WriteLine($"Error fetching storage quotas: {error.Message}");
             return [];
         }
     }
